Изменен тип данных скидки с Double на Integer, создан утилитный класс для применения скидки

+++ src/main/java/org/philosophy/carwashing/dto/DiscountableDto.java
@@ -14,6 +14,6 @@ import lombok.*;
 public abstract class DiscountableDto {
 
     protected Boolean hasDiscount;
-    protected Double discountAmount;
+    protected Integer discountAmount;
 
 }
